+++ AssemblyInfo.cs
@@ -1,3 +1,4 @@
 using System.Runtime.CompilerServices;
 
 [assembly: InternalsVisibleTo("PokeChat.Tests")]
+[assembly: InternalsVisibleTo("PokeChat.Api")]