+++ NLP/IPosTagger.cs
@@ -3,4 +3,5 @@ namespace PokeChat.NLP;
 public interface IPosTagger
 {
     List<PosTag> Tag(List<string> tokens);
+    bool IsKnownWord(string word);
 }